src/user_service/requirements.txt
fastapi
requests
uvicorn[standard]
psycopg2-binary
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-auth
httpx
pinecone